🚚 rename: 나의 팀 불러오기 및 삭제 api 폴더를 common으로 이동 #50
